Fix categories and checkbox

+++ OsmAnd/src/net/osmand/plus/ContextMenuAdapter.java
@@ -214,7 +214,7 @@ public class ContextMenuAdapter {
 						public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
 							OnContextMenuClick ca = getClickAdapter(position);
 							if(ca != null) {
-								ca.onContextMenuClick(getItemId(position), position, isChecked);
+								ca.onContextMenuClick((int) getItemId(position), position, isChecked);
 							}
 						}
 					});
